+++ .gitignore
@@ -77,5 +77,5 @@ celerybeat-schedule.*
 # Internal files
 .env
 .aws/
-payload
+s3_downloads/
 tmp/
