+++ site/mex_invenio/records/api.py
@@ -9,7 +9,7 @@ from invenio_records.dumpers.relations import RelationDumperExt
 from invenio_records_resources.records.dumpers import CustomFieldsDumperExt
 from invenio_records_resources.records.systemfields import IndexField
 
-from mex_invenio.custom_search import MexDumper
+from mex_invenio.services.search import MexDumper
 from mex_invenio.systemfields import IndexField as MexIndexField
