+++ site/mex_invenio/search/__init__.py
@@ -1,0 +1,1 @@
+"""Search package for MEX Invenio."""
