+++ site/mex_invenio/views/__init__.py
@@ -1,0 +1,1 @@
+"""Views package for MEX Invenio."""
