+++ site/setup.cfg
@@ -7,7 +7,7 @@ tests =
 
 [options.entry_points]
 invenio_base.blueprints =
-    mex_invenio_views = mex_invenio.views:create_blueprint
+    mex_invenio_views = mex_invenio.views.blueprints:create_blueprint
 invenio_assets.webpack =
     mex_invenio_theme = mex_invenio.webpack:theme
 invenio_search.mappings =
